feat(media): switch to human readable size
